requirements.txt
Django==2.2.6
django-crispy-forms==1.7.2
djangorestframework==3.10.3
psycopg2-binary==2.8.3
python-dotenv==0.10.3
pytz==2019.3
sqlparse==0.3.0
